+++ SuperFoobar/Character filer/Character.h
@@ -9,14 +9,14 @@ class Character
 public:
 	Character();
 	virtual ~Character() = default;
-	Character(float, float, float, float)
+	Character(int, int, int, int)
 		: position {new Cluster(x_pos, y_pos, height, width) } {}
-	void move_x(float);
-	void move_y(float);
-	float get_x_pos();
-	float get_y_pos();
-	void set_x_pos(float);
-	void set_y_pos(float);
+	void move_x(int);
+	void move_y(int);
+	int get_x_pos();
+	int get_y_pos();
+	void set_x_pos(int);
+	void set_y_pos(int);
 
 	void die();
